move VirtualKeyboard to subdirectory
