+++ libs/ui/src/index.ts
@@ -93,7 +93,7 @@ export * from './usb_drive_image';
 export * from './power_down_button';
 export * from './verify_ballot_image';
 export * from './visual_mode_disabled_overlay';
-export * from './virtual_keyboard';
+export * from './virtual_keyboard/virtual_keyboard';
 export * from './voter_contest_summary';
 export * from './with_scroll_buttons';
 export * from './search_select';
