+++ libs/ui/src/virtual_keyboard/common.ts
@@ -1,0 +1,19 @@
+import { DefaultTheme } from 'styled-components';
+
+export interface Key {
+  audioLanguageOverride?: string;
+  renderAudioString: () => React.ReactNode;
+  /** @defaultvalue () => {@link value} */
+  renderLabel?: () => React.ReactNode;
+  value: string;
+}
+
+/* istanbul ignore next */
+export function getBorderWidthRem(p: { theme: DefaultTheme }): number {
+  switch (p.theme.sizeMode) {
+    case 'touchExtraLarge':
+      return p.theme.sizes.bordersRem.hairline;
+    default:
+      return p.theme.sizes.bordersRem.thin;
+  }
+}
